feat: move to rest package, fix translations
