+++ src/main/java/sk/iway/iwcm/stat/rest/BrowserIdentifierMigrationRestController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
-import sk.iway.iwcm.stat.BrowserIdentifierMigrationService;
 
 @RestController
 @RequiredArgsConstructor
