+++ src/main/java/sk/iway/iwcm/system/Modules.java
@@ -789,7 +789,7 @@ public class Modules
 		sub.setShowInLeftMenu(true);
 		sub.setGroup("config");
 		sub.setMenuIcon("refresh");
-		ModuleInfo browserMigration = new ModuleInfo().setMenuOrder(7301);
+		/*ModuleInfo browserMigration = new ModuleInfo().setMenuOrder(7301);
 		browserMigration.setNameKey("stat.browser-migration.title");
 		browserMigration.setPath("/admin/v9/settings/stat-browser-migration/");
 		browserMigration.setWjVersions("BPECDIM");
@@ -799,7 +799,7 @@ public class Modules
 		browserMigration.setShowInLeftMenu(false);
 		browserMigration.setGroup("config");
 		sub.addSubmenu(browserMigration);
-		sub.setHideSubmenu(true);
+		sub.setHideSubmenu(true);*/
 		loadModule(sub);
 
 		// Embedding-chunks
