+++ src/main/webapp/WEB-INF/classes/text-webjet9.properties
@@ -3137,3 +3137,5 @@ stat.browser-migration.key-count=zlúčení identifikátorov prehliadača
 stat.browser-migration.table-count=tabuliek
 stat.browser-migration.source=Pôvodná hodnota
 stat.browser-migration.target=Nová hodnota
+stat.browser-migration.warning=Databáza štatistík ešte nebola skonvertovaná na nový formát identifikátorov prehliadačov.
+stat.browser-migration.warning-start=Spustiť konverziu
