+++ src/main/webapp/WEB-INF/classes/text_cz-webjet9.properties
@@ -3137,3 +3137,5 @@ stat.browser-migration.key-count=sloučení identifikátorů prohlížeče
 stat.browser-migration.table-count=tabulek
 stat.browser-migration.source=Původní hodnota
 stat.browser-migration.target=Nová hodnota
+stat.browser-migration.warning=Databáze statistik ještě nebyla převedena na nový formát identifikátorů prohlížečů.
+stat.browser-migration.warning-start=Spustit konverzi
