+++ src/main/webapp/WEB-INF/classes/text_en-webjet9.properties
@@ -3137,3 +3137,5 @@ stat.browser-migration.key-count=browser identifier merges
 stat.browser-migration.table-count=tables
 stat.browser-migration.source=Original value
 stat.browser-migration.target=New value
+stat.browser-migration.warning=The statistics database has not yet been converted to the new browser identifier format.
+stat.browser-migration.warning-start=Start conversion
